scraping_agent_firecrawl/requirements.txt
firecrawl-py
openai
dotenv
pandas
openpyxl
Flask